+++ StabilityMatrix.Avalonia/DesignData/DesignData.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Net.Http;
 using System.Text;
 using AvaloniaEdit.Utils;
+using DynamicData;
 using DynamicData.Binding;
 using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
@@ -175,6 +176,55 @@ public static class DesignData
         );
         InstallerViewModel.SelectedPackage = InstallerViewModel.AvailablePackages[0];
         InstallerViewModel.ReleaseNotes = "## Release Notes\nThis is a test release note.";
+
+        ObservableCacheEx.AddOrUpdate(
+            CheckpointsPageViewModel.CheckpointFoldersCache,
+            new CheckpointFolder[]
+            {
+                new(settingsManager, downloadService, modelFinder, notificationService)
+                {
+                    DirectoryPath = "Models/StableDiffusion",
+                    DisplayedCheckpointFiles = new ObservableCollectionExtended<CheckpointFile>()
+                    {
+                        new()
+                        {
+                            FilePath = "~/Models/StableDiffusion/electricity-light.safetensors",
+                            Title = "Auroral Background",
+                            PreviewImagePath =
+                                "https://image.civitai.com/xG1nkqKTMzGDvpLrqFT7WA/"
+                                + "78fd2a0a-42b6-42b0-9815-81cb11bb3d05/00009-2423234823.jpeg",
+                            ConnectedModel = new ConnectedModelInfo
+                            {
+                                VersionName = "Lightning Auroral",
+                                BaseModel = "SD 1.5",
+                                ModelName = "Auroral Background",
+                                ModelType = CivitModelType.Model,
+                                FileMetadata = new CivitFileMetadata
+                                {
+                                    Format = CivitModelFormat.SafeTensor,
+                                    Fp = CivitModelFpType.fp16,
+                                    Size = CivitModelSize.pruned,
+                                }
+                            }
+                        },
+                        new()
+                        {
+                            FilePath = "~/Models/Lora/model.safetensors",
+                            Title = "Some model"
+                        },
+                    },
+                },
+                new(settingsManager, downloadService, modelFinder, notificationService)
+                {
+                    Title = "Lora",
+                    DirectoryPath = "Packages/Lora",
+                    DisplayedCheckpointFiles = new ObservableCollectionExtended<CheckpointFile>
+                    {
+                        new() { FilePath = "~/Models/Lora/lora_v2.pt", Title = "Best Lora v2", }
+                    }
+                }
+            }
+        );
 
         /*// Checkpoints page
         CheckpointsPageViewModel.CheckpointFolders =
@@ -621,8 +671,8 @@ The gallery images are often inpainted, but you will get something very similar 
         get
         {
             var list = new CompletionList { IsFiltering = true };
-            list.CompletionData.AddRange(SampleCompletionData);
-            list.FilteredCompletionData.AddRange(list.CompletionData);
+            ExtensionMethods.AddRange(list.CompletionData, SampleCompletionData);
+            ExtensionMethods.AddRange(list.FilteredCompletionData, list.CompletionData);
             list.SelectItem("te", true);
             return list;
         }